// frontend/src/api/generated/models/intensityTargetType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export type IntensityTargetType = typeof IntensityTargetType[keyof typeof IntensityTargetType];


export const IntensityTargetType = {
  Free: 'Free',
  HeartRateZone: 'HeartRateZone',
  Pace: 'Pace',
  Rpe: 'Rpe',
  Power: 'Power',
} as const;